Clean up Motives
